fix: use the fully-merged names index instead of a local shadow copy
Codex review, fresh evidence (ninth round finding, comment at line 900):
the eighth round's default-translation fix kept its own local "first name
seen for this qualname" shadow inside _index_template_param_defaults's
walk, updated only at a qualname's very first sighting -- which silently
went stale whenever an INTERMEDIATE redeclaration was the one that
actually named a previously-unnamed parameter: `template<class, class>
struct A;` (both unnamed), then `template<class X, class Y> struct A;` (a
redeclaration that names them), then `template<class T, class U=T> struct
A {...};` (a further redeclaration that adds a default). The real,
fully-merged names index correctly resolves to ["X", "Y"], but the stale
local shadow still held [None, None] from the very first sighting, so the
added default's translation target was falsy and the raw, untranslated
text ("T") was kept -- reproducing the identical unresolvable-base
failure.

Verified with a hand-built three-declaration AST fixture reproducing the
exact scenario -- confirmed the new test fails against the pre-fix code
with the untranslated "T" instead of the expected "X", and passes after
the fix.

Fixed by reusing _index_template_param_names's own already-correct,
fully-merged result directly as the translation target (computed once up
front) instead of re-deriving a second, independently-tracked copy that
can diverge from it. Confirmed the third through eighth-round regression
tests all still pass unaffected.

Also splits tests/test_dumper_clang_vtable.py (which crossed the
AI-readiness 2000-line hard cap after this round's new test) into itself
(own-vtable/base-lookup/specialization-spelling coverage) plus a new
sibling tests/test_dumper_clang_vtable_redecl.py (class-template
redeclaration-identity coverage, rounds 3-9), sharing their hand-built AST
node builders via a new tests/_dumper_clang_vtable_helpers.py leaf module
-- mirroring the existing tests/_libabigail.py /
tests/_comparability_gate_helpers.py shared-helper pattern. Test-only
split, no production code moved, exact same 70 vtable-module tests still
pass split across the two files.

Full fast suite: 24683 passed, 0 failed. mypy/ai-readiness clean.
