+++ tests/test_dumper_clang_vtable.py
@@ -21,6 +21,14 @@ emit surface is exercised without clang installed). Every shape here was
 first verified against a REAL ``clang++ -Xclang -ast-dump=json`` run before
 being reduced to a fixture -- see ``dumper_clang_vtable.py``'s own docstring
 for the exact empirical findings this module's design rests on.
+
+Class-template REDECLARATION identity (kinds/defaults/names merging across
+a ``previousDecl`` chain, and the ambiguity guard for two unrelated
+declarations sharing a bare qualname) lives in the sibling
+``test_dumper_clang_vtable_redecl.py`` instead -- split out once this file
+crossed the AI-readiness 2000-line hard cap (``tests/CLAUDE.md``'s
+file-size convention). Both files share their hand-built node builders via
+``tests._dumper_clang_vtable_helpers``.
 """
 
 from __future__ import annotations
@@ -32,67 +40,16 @@ from abicheck.dumper_clang_vtable import (
     _normalize_param_type,
     _top_level_param_list_close,
 )
-
-
-def _tu(*inner: dict) -> dict:
-    return {"kind": "TranslationUnitDecl", "inner": list(inner)}
-
-
-def _record(name: str, *inner: dict, bases: list[dict] | None = None) -> dict:
-    node = {
-        "kind": "CXXRecordDecl",
-        "name": name,
-        "tagUsed": "struct",
-        "loc": {"file": "include/foo.h", "line": 1},
-        "completeDefinition": True,
-        "inner": list(inner),
-    }
-    if bases:
-        node["bases"] = bases
-    return node
-
-
-def _base(qualtype: str, *, is_virtual: bool = False) -> dict:
-    return {"type": {"qualType": qualtype}, "access": "public", "isVirtual": is_virtual}
-
-
-def _method(
-    name: str,
-    mangled: str,
-    *,
-    virtual: bool = False,
-    override_attr: bool = False,
-    params: list[str] | None = None,
-    is_const: bool = False,
-) -> dict:
-    qual = f"void ({', '.join(params or [])})" + (" const" if is_const else "")
-    inner = [{"kind": "ParmVarDecl", "type": {"qualType": p}} for p in (params or [])]
-    if override_attr:
-        inner.append({"kind": "OverrideAttr"})
-    node: dict = {
-        "kind": "CXXMethodDecl",
-        "name": name,
-        "mangledName": mangled,
-        "type": {"qualType": qual},
-        "inner": inner,
-    }
-    if virtual:
-        node["virtual"] = True
-    return node
-
-
-def _dtor(mangled: str, *, virtual: bool = False, implicit: bool = False) -> dict:
-    node: dict = {"kind": "CXXDestructorDecl", "mangledName": mangled}
-    if virtual:
-        node["virtual"] = True
-    if implicit:
-        node["isImplicit"] = True
-    return node
-
-
-def _types(root: dict) -> dict[str, object]:
-    return {t.name: t for t in _ClangAstParser(root, set(), set()).parse_types()}
-
+from tests._dumper_clang_vtable_helpers import (
+    _base,
+    _dtor,
+    _forward_specialization,
+    _method,
+    _record,
+    _specialization,
+    _tu,
+    _types,
+)
 
 # ── primary vtable, no inheritance ────────────────────────────────────────
 
@@ -796,21 +753,6 @@ def test_ref_qualifier_mismatch_survives_trailing_exception_spec() -> None:
 # ── record (Codex review, fresh evidence, P1) ──────────────────────────
 
 
-def _specialization(name: str, *inner: dict, type_args: list[str]) -> dict:
-    """A ``ClassTemplateSpecializationDecl`` node, mirroring real clang
-    output: one ``TemplateArgument`` child per type argument (each carrying
-    only the ``type.qualType`` this module's spelling-reconstruction
-    reads), followed by *inner*'s own children.
-    """
-    args = [{"kind": "TemplateArgument", "type": {"qualType": t}} for t in type_args]
-    return {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": name,
-        "completeDefinition": True,
-        "inner": [*args, *inner],
-    }
-
-
 def test_base_resolved_via_concrete_template_specialization() -> None:
     """A base that is a CONCRETE template specialization (``struct D :
     A<int> {...};``) used to be entirely unresolvable: clang emits the
@@ -901,18 +843,6 @@ def test_specialization_base_override_with_no_keyword_replaces_slot_in_place() -
     # matches D.bases == ["A<int>"] and produces a false
     # TYPE_VTABLE_CHANGED once the base above is resolved at all.
     assert funcs["_ZN1AIiE1fEi"].name == "A<int>::f"
-
-
-def _forward_specialization(name: str, *, type_args: list[str]) -> dict:
-    """A forward-declared explicit specialization node -- no
-    ``completeDefinition``, no member children -- sharing the same spelling
-    a later complete definition would."""
-    args = [{"kind": "TemplateArgument", "type": {"qualType": t}} for t in type_args]
-    return {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": name,
-        "inner": args,
-    }
 
 
 def test_specialization_forward_declaration_does_not_shadow_complete_definition() -> (
@@ -1514,448 +1444,4 @@ def test_nested_specialization_indexes_under_outer_specialization_qualname() -> 
     )
     types = _types(root)
     assert types["D"].vtable == ["_ZN5OuterIiE1AIdE1fEv"]
-    assert types["D"].vptr_offset_bits == 0
-
-
-def test_conflicting_nested_template_defaults_across_sibling_specializations_stay_ambiguous() -> (
-    None
-):
-    """Codex review, fresh evidence (P1, third round): when TWO different
-    explicit outer specializations each define their OWN same-named nested
-    member template with DIFFERING defaults (``Outer<int>``'s ``template
-    <class U=int> struct A`` vs. ``Outer<double>``'s ``template<class
-    U=double> struct A``), both register under the identical BARE qualname
-    ``"A"`` -- neither ``_index_template_param_kinds`` nor its two
-    siblings ever extend scope through a ``ClassTemplateSpecializationDecl``.
-    First-registration-wins previously trusted whichever was visited
-    first, silently borrowing the WRONG default for the other -- confirmed
-    end-to-end this left the base unresolvable and a real virtual-method
-    addition on the derived class completely undetected (`NO_CHANGE`).
-    Fixed by treating a conflicting second registration as genuinely
-    ambiguous and dropping it entirely, rather than trusting either
-    candidate.
-    """
-    outer_int_spec = {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": "Outer",
-        "completeDefinition": True,
-        "inner": [
-            {"kind": "TemplateArgument", "type": {"qualType": "int"}},
-            {
-                "kind": "ClassTemplateDecl",
-                "name": "A",
-                "inner": [
-                    {
-                        "kind": "TemplateTypeParmDecl",
-                        "name": "U",
-                        "defaultArg": {
-                            "kind": "TemplateArgument",
-                            "type": {"qualType": "int"},
-                        },
-                    },
-                    _record("A"),
-                ],
-            },
-        ],
-    }
-    outer_double_spec = {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": "Outer",
-        "completeDefinition": True,
-        "inner": [
-            {"kind": "TemplateArgument", "type": {"qualType": "double"}},
-            {
-                "kind": "ClassTemplateDecl",
-                "name": "A",
-                "inner": [
-                    {
-                        "kind": "TemplateTypeParmDecl",
-                        "name": "U",
-                        "defaultArg": {
-                            "kind": "TemplateArgument",
-                            "type": {"qualType": "double"},
-                        },
-                    },
-                    _record("A"),
-                    _specialization(
-                        "A",
-                        {
-                            "kind": "CXXMethodDecl",
-                            "name": "f",
-                            "mangledName": "_ZN5OuterIdE1AIdE1fEv",
-                            "type": {"qualType": "void ()"},
-                            "virtual": True,
-                        },
-                        type_args=["double"],
-                    ),
-                ],
-            },
-        ],
-    }
-    root = _tu(
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "Outer",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                _record("Outer"),
-            ],
-        },
-        outer_int_spec,
-        outer_double_spec,
-        _record("D", bases=[_base("Outer<double>::A<>")]),
-    )
-    from abicheck.dumper_clang_vtable import _index_template_param_defaults
-
-    # Both nested "A"s are type-only parameters (the KINDS list -- type vs.
-    # non-type -- is identical, [None], for both), so only the DEFAULTS
-    # index (where the two genuinely differ, "int" vs. "double") is the one
-    # that must detect the conflict and drop the ambiguous entry.
-    assert "A" not in _index_template_param_defaults(root)
-    # Safe degradation, not a crash or a wrong resolution: the base stays
-    # unresolvable (same accepted false negative every other unresolvable-
-    # base shape in this module degrades to), never a fabricated slot.
-    types = _types(root)
-    assert types["D"].vtable == []
-
-
-def test_legal_redeclaration_with_renamed_params_still_resolves_dependent_default() -> (
-    None
-):
-    """Codex review, fresh evidence (fourth round): the third round's
-    conflicting-registration-is-ambiguous fix was too broad -- an ordinary,
-    LEGAL C++ redeclaration of the SAME template with renamed parameters
-    (`template<class T, class U=T> struct A;` followed by `template<class
-    X, class Y> struct A {...};`) must not be treated as a genuine
-    conflict just because the literal parameter NAMES differ across the
-    two declarations. Clang always spells a dependent default's inherited
-    text using the ORIGINAL declaration's parameter name (`"T"`, never
-    renamed to `"X"`/`"Y"` on the later declaration, confirmed
-    empirically), so dropping the NAMES metadata here broke dependent-
-    default substitution for this ordinary shape -- distinguished from a
-    genuine conflict via clang's own `previousDecl` link (present on every
-    legal redeclaration, absent between two genuinely unrelated
-    declarations that coincidentally share a bare qualname).
-    """
-    root = _tu(
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                {
-                    "kind": "TemplateTypeParmDecl",
-                    "name": "U",
-                    "defaultArg": {
-                        "kind": "TemplateArgument",
-                        "type": {"qualType": "T"},
-                    },
-                },
-            ],
-        },
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x2",
-            "previousDecl": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "X"},
-                {"kind": "TemplateTypeParmDecl", "name": "Y"},
-                _record("A"),
-                _specialization(
-                    "A",
-                    {
-                        "kind": "CXXMethodDecl",
-                        "name": "f",
-                        "mangledName": "_ZN1AIddE1fEv",
-                        "type": {"qualType": "void ()"},
-                        "virtual": True,
-                    },
-                    type_args=["double", "double"],
-                ),
-            ],
-        },
-        _record("D", bases=[_base("A<double>")]),
-    )
-    types = _types(root)
-    assert types["D"].vtable == ["_ZN1AIddE1fEv"]
-    assert types["D"].vptr_offset_bits == 0
-
-
-def test_unrelated_coincidentally_equal_registration_does_not_corrupt_redecl_identity() -> (
-    None
-):
-    """Codex review, fresh evidence (fifth round): the fourth round's
-    ``previousDecl``-based redeclaration check was itself incomplete.
-    ``_register_template_param_metadata`` also treated a bare *value*
-    match (``existing == value``) as license to advance the tracked
-    ``node_ids[qualname]`` -- but two genuinely UNRELATED nested member
-    templates (``Outer<int>``'s and ``Outer<double>``'s own, neither ever
-    redeclared) can trivially have byte-identical kinds/defaults/names by
-    coincidence before either is redeclared: both are plain ``template
-    <class T, class U=T> struct A;``. Advancing ``node_ids["A"]`` to the
-    SECOND (unrelated) node's id there corrupted the chain for the FIRST
-    entity's own later, legal out-of-class redeclaration with renamed
-    parameters: its real ``previousDecl`` (pointing at the first node)
-    then mismatched the corrupted tracked id and was wrongly dropped as
-    ambiguous, losing the DEFAULTS metadata for "A" entirely and leaving
-    ``Outer<int>::A<double>``'s dependent default unsubstituted -- the
-    specialization was indexed as ``Outer<int>::A<double, double>``
-    instead of the base's actual ``Outer<int>::A<double>``, leaving the
-    inherited vtable unresolvable and a real added virtual method
-    undetected. Fixed by only ever advancing ``node_ids`` on a CONFIRMED
-    same-entity link (the first registration, or a ``previousDecl``
-    match) -- a coincidental value match no longer reassigns whose id is
-    being tracked.
-    """
-    outer_int_spec = {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": "Outer",
-        "completeDefinition": True,
-        "inner": [
-            {"kind": "TemplateArgument", "type": {"qualType": "int"}},
-            {
-                "kind": "ClassTemplateDecl",
-                "name": "A",
-                "id": "0x1",
-                "inner": [
-                    {"kind": "TemplateTypeParmDecl", "name": "T"},
-                    {
-                        "kind": "TemplateTypeParmDecl",
-                        "name": "U",
-                        "defaultArg": {
-                            "kind": "TemplateArgument",
-                            "type": {"qualType": "T"},
-                        },
-                    },
-                ],
-            },
-        ],
-    }
-    outer_double_spec = {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": "Outer",
-        "completeDefinition": True,
-        "inner": [
-            {"kind": "TemplateArgument", "type": {"qualType": "double"}},
-            {
-                "kind": "ClassTemplateDecl",
-                "name": "A",
-                "id": "0x2",
-                "inner": [
-                    {"kind": "TemplateTypeParmDecl", "name": "T"},
-                    {
-                        "kind": "TemplateTypeParmDecl",
-                        "name": "U",
-                        "defaultArg": {
-                            "kind": "TemplateArgument",
-                            "type": {"qualType": "T"},
-                        },
-                    },
-                ],
-            },
-        ],
-    }
-    # Out-of-class definition of Outer<int>::A, using RENAMED parameters --
-    # a legal redeclaration of the FIRST entity (id "0x1"), never the
-    # second (unrelated) one.
-    outer_int_spec_redef = {
-        "kind": "ClassTemplateSpecializationDecl",
-        "name": "Outer",
-        "completeDefinition": True,
-        "inner": [
-            {"kind": "TemplateArgument", "type": {"qualType": "int"}},
-            {
-                "kind": "ClassTemplateDecl",
-                "name": "A",
-                "id": "0x3",
-                "previousDecl": "0x1",
-                "inner": [
-                    {"kind": "TemplateTypeParmDecl", "name": "X"},
-                    {
-                        "kind": "TemplateTypeParmDecl",
-                        "name": "Y",
-                        "defaultArg": {
-                            "kind": "TemplateArgument",
-                            "type": {"qualType": "X"},
-                        },
-                    },
-                    _record("A"),
-                    _specialization(
-                        "A",
-                        {
-                            "kind": "CXXMethodDecl",
-                            "name": "f",
-                            "mangledName": "_ZN5OuterIiE1AIddE1fEv",
-                            "type": {"qualType": "void ()"},
-                            "virtual": True,
-                        },
-                        type_args=["double", "double"],
-                    ),
-                ],
-            },
-        ],
-    }
-    root = _tu(
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "Outer",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                _record("Outer"),
-            ],
-        },
-        outer_int_spec,
-        outer_double_spec,
-        outer_int_spec_redef,
-        _record("D", bases=[_base("Outer<int>::A<double>")]),
-    )
-    from abicheck.dumper_clang_vtable import _index_template_param_defaults
-
-    # The redeclaration's own previousDecl link must still resolve
-    # correctly despite the unrelated, coincidentally-equal middle
-    # registration -- "A" stays in the index (not dropped as ambiguous),
-    # keeping the FIRST declaration's defaults (dependent default spelled
-    # with the ORIGINAL parameter name "T", not the redeclaration's "X").
-    assert _index_template_param_defaults(root).get("A") == [None, "T"]
-    types = _types(root)
-    assert types["D"].vtable == ["_ZN5OuterIiE1AIddE1fEv"]
-    assert types["D"].vptr_offset_bits == 0
-
-
-def test_default_added_by_legal_redeclaration_still_resolves_dependent_default() -> None:
-    """Codex review, fresh evidence (sixth round): a CONFIRMED redeclaration
-    (linked via ``previousDecl``) previously kept the tracked metadata frozen
-    at the FIRST declaration's value unconditionally -- correct for a
-    renamed-parameter redeclaration (the fourth round's own case, where the
-    original names/defaults must win), but wrong when the later declaration
-    legally ADDS a default the first one never had:
-    ``template<class T, class U> struct A;`` followed by ``template<class T,
-    class U=T> struct A {...};`` is one C++ entity whose effective default
-    for ``U`` only becomes visible on the SECOND declaration. Dropping it
-    left ``_index_template_param_defaults`` with ``[None, None]`` for "A",
-    so dependent-default substitution couldn't trim the trailing argument --
-    mis-indexing ``A<double>`` as ``A<double, double>``, unable to match the
-    base's actual trimmed spelling, leaving the inherited vtable invisible.
-    Fixed by positionally MERGING the confirmed redeclaration's value into
-    the tracked one (a position the tracked value has no data for adopts the
-    new declaration's value there) rather than keeping it frozen.
-    """
-    root = _tu(
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                {"kind": "TemplateTypeParmDecl", "name": "U"},
-            ],
-        },
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x2",
-            "previousDecl": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                {
-                    "kind": "TemplateTypeParmDecl",
-                    "name": "U",
-                    "defaultArg": {
-                        "kind": "TemplateArgument",
-                        "type": {"qualType": "T"},
-                    },
-                },
-                _record("A"),
-                _specialization(
-                    "A",
-                    {
-                        "kind": "CXXMethodDecl",
-                        "name": "f",
-                        "mangledName": "_ZN1AIddE1fEv",
-                        "type": {"qualType": "void ()"},
-                        "virtual": True,
-                    },
-                    type_args=["double", "double"],
-                ),
-            ],
-        },
-        _record("D", bases=[_base("A<double>")]),
-    )
-    from abicheck.dumper_clang_vtable import _index_template_param_defaults
-
-    # "A" now carries the default the SECOND declaration added, letting
-    # dependent-default substitution trim "A<double, double>" down to the
-    # base's actual spelling "A<double>".
-    assert _index_template_param_defaults(root).get("A") == [None, "T"]
-    types = _types(root)
-    assert types["D"].vtable == ["_ZN1AIddE1fEv"]
-    assert types["D"].vptr_offset_bits == 0
-
-
-def test_added_default_with_renamed_params_translates_through_positions() -> None:
-    """Codex review, fresh evidence (eighth round): the sixth round's merge
-    logic adopted a newly-added default's raw text as-is, but when the
-    redeclaration that adds the default ALSO renames its parameters (
-    ``template<class T, class U> struct A;`` followed by ``template<class
-    X, class Y=X> struct A {...};``), that raw text ("X") names one of the
-    REDECLARATION's own (renamed) parameters, not the tracked/original
-    names ("T"/"U") :func:`_specialization_spelling` looks the dependent
-    reference up against -- so the substitution silently failed to find
-    "X" in the tracked names index, leaving the trailing argument
-    untrimmed. Fixed by translating a newly adopted default's dependent
-    reference through the CONTRIBUTING declaration's own positional name
-    list into the FIRST declaration's name at that same position before
-    merging it in.
-    """
-    root = _tu(
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "T"},
-                {"kind": "TemplateTypeParmDecl", "name": "U"},
-            ],
-        },
-        {
-            "kind": "ClassTemplateDecl",
-            "name": "A",
-            "id": "0x2",
-            "previousDecl": "0x1",
-            "inner": [
-                {"kind": "TemplateTypeParmDecl", "name": "X"},
-                {
-                    "kind": "TemplateTypeParmDecl",
-                    "name": "Y",
-                    "defaultArg": {
-                        "kind": "TemplateArgument",
-                        "type": {"qualType": "X"},
-                    },
-                },
-                _record("A"),
-                _specialization(
-                    "A",
-                    {
-                        "kind": "CXXMethodDecl",
-                        "name": "f",
-                        "mangledName": "_ZN1AIddE1fEv",
-                        "type": {"qualType": "void ()"},
-                        "virtual": True,
-                    },
-                    type_args=["double", "double"],
-                ),
-            ],
-        },
-        _record("D", bases=[_base("A<double>")]),
-    )
-    from abicheck.dumper_clang_vtable import _index_template_param_defaults
-
-    # The added default is translated to the FIRST declaration's own
-    # parameter name ("T"), matching the tracked names index used by
-    # _specialization_spelling's dependent-default substitution.
-    assert _index_template_param_defaults(root).get("A") == [None, "T"]
-    types = _types(root)
-    assert types["D"].vtable == ["_ZN1AIddE1fEv"]
     assert types["D"].vptr_offset_bits == 0
